We almost never use post-formats
So far I haven't seen a use case for ['post-formats'](https://codex.wordpress.org/Post_Formats) in our projects. We find ourselves always needing to remove this line to avoid unnecessary clutter having an empty "Post Formats" metabox widget in the Posts edit backend.

Maybe it does makes sense to remove this line from the base Gesso-WP theme? If in a future project we find a use case for them, we can always put it back for that project specifically.